+++ R/rextendr_document.R
@@ -5,15 +5,17 @@
 #' made in the Rust code. It is a wrapper for [devtools::document()], and it
 #' executes `extendr`-specific routines before calling [devtools::document()].
 #' Specifically, it ensures that Rust code is recompiled (when necessary) and that
-#' up-to-date R wrappers are generated before re-generating the package documentation.
+#' up-to-date R wrappers are generated before regenerating the package documentation.
 #' @inheritParams devtools::document
 #' @return No return value, called for side effects.
 #' @export
-document <- function(pkg = ".", quiet = getOption("usethis.quiet", FALSE), roclets = NULL) {
+document <- function(pkg = ".", quiet = FALSE, roclets = NULL) {
   try_save_all(quiet = quiet)
+
+  withr::local_envvar(devtools::r_env_vars())
 
   register_extendr(path = pkg, quiet = quiet)
 
   rlang::check_installed("devtools")
-  devtools::document(pkg = pkg, roclets = roclets, quiet = FALSE)
+  devtools::document(pkg = pkg, roclets = roclets, quiet = quiet)
 }