+++ R/zzz.R
@@ -1,5 +1,4 @@
 .onLoad <- function(...) {
-
   # register the extendr knitr chunk engine if knitr is available
   if (requireNamespace("knitr", quietly = TRUE)) {
     knitr::knit_engines$set(
@@ -13,8 +12,6 @@
   # NULL values are present for reference and may later be replaced
   # by concrete values
 
-  git_ref <- list(git = "https://github.com/extendr/extendr")
-
   rextendr_opts <- list(
     # Controls default Rust toolchain; NULL corresponds to system's default
     rextendr.toolchain = NULL,
@@ -23,12 +20,15 @@
     # Overrides Rust dependencies; mainly used for development
     rextendr.patch.crates_io = NULL, # most recent extendr crates on crates.io
     #    rextendr.patch.crates_io = list(  # most recent extendr crates on github
-    #      `extendr-api` = git_ref
+    #      `extendr-api` = list(git = "https://github.com/extendr/extendr")
     #    ),
 
     # Version of 'extendr_api' to be used
     rextendr.extendr_deps = list(
       `extendr-api` = "*"
+    ),
+    rextendr.extendr_dev_deps = list(
+      `extendr-api` = list(git = "https://github.com/extendr/extendr")
     )
   )